Try using f32 internally for udim.
It's kind of annoying to support both i32 & f32 (and I need floats for
lerping) in udim. Either we can have udim<i32> and then temporarily go
to a dimensionless f32, or we just do everything in a udim<f32>.
Of course handling pixels with floats is disgusting so I'll probably go
back to i32 at some point.